Update notify email script to show changes
