+++ package.json
@@ -79,7 +79,7 @@
     "storybook": "start-storybook -p 9009 -s public",
     "build-storybook": "build-storybook -s public",
     "codegen": "graphql-codegen --config codegen.js",
-    "notify-email": "env-cmd -e prod -r config/.cmdrc.json node ./notify-email.js",
+    "notify-email": "env-cmd -e prod -r config/.cmdrc.json node cli-scripts/notify-email.js",
     "deploy-prod:do-not-use-directly": "npm run build:prod && BUCKET=evergreen.spruce npm run deploy; npm run notify-email; npm run upload-source-maps;",
     "upload-source-maps": "env-cmd -e prod -r config/.cmdrc.json node ./upload-bugsnag-sourcemaps"
   },
